compositor|refactor: Give the target schemas their own module
`TargetEntrySchema` and the token-mapping vocabulary now live in `target-schemas.ts`, leaving `descriptor-schemas.ts` holding the kinds, sources, and tiers that the catalog and config read alongside the plan.

`scalar-schemas.ts` gains the test module it lacked, taking the `IdSchema` cases that had been filed under the descriptors.
