+++ packages/compositor/src/schemas/__tests__/scalar-schemas.unit.test.ts
@@ -1,0 +1,10 @@
+import { describe, expect, it } from 'vitest';
+
+import { IdSchema } from '../scalar-schemas.ts';
+import { findIssuePaths } from '../test-utils/findIssuePaths.ts';
+
+describe('IdSchema', () => {
+  it('if the id is empty, rejects it at the root', () => {
+    expect(findIssuePaths(IdSchema, '')).toStrictEqual([[]]);
+  });
+});
