+++ packages/compositor/src/schemas/descriptor-schemas.ts
@@ -1,4 +1,4 @@
-/** Descriptors: the kinds, sources, tiers, and targets a plan is computed over. */
+/** Descriptors: the kinds, sources, and tiers a plan is computed over. */
 
 import { z } from 'zod';
 
@@ -55,55 +55,7 @@ export const TierDescriptorSchema = z
   })
   .meta({ id: 'TierDescriptor' });
 
-/** One canonical-name-to-rendered-name pair within a token kind's mapping. */
-export const TokenMappingEntrySchema = z
-  .object({
-    from: z.string(),
-    to: z.string(),
-  })
-  .meta({ id: 'TokenMappingEntry' });
-
-/**
- * How one token kind's canonical names render for a target.
- *
- * Token kinds are declared as consumer data, so a tool-name mapping and any other name-rewriting vocabulary share this
- * one shape.
- */
-export const TokenMappingSchema = z
-  .object({
-    kindId: IdSchema,
-    entries: z.array(TokenMappingEntrySchema),
-  })
-  .meta({ id: 'TokenMapping' });
-
-/** One named value a target's transforms are parameterized by. */
-export const TargetVariableSchema = z
-  .object({
-    name: z.string(),
-    value: z.string(),
-  })
-  .meta({ id: 'TargetVariable' });
-
-/**
- * One destination the plan renders content for, with the metadata a consumer can query.
- *
- * `tokenMappings` and `variables` are arrays of pairs, so each has a sort key the determinism guarantee can name.
- */
-export const TargetEntrySchema = z
-  .object({
-    id: IdSchema,
-    label: z.string(),
-    root: z.string(),
-    tokenMappings: z.array(TokenMappingSchema),
-    variables: z.array(TargetVariableSchema),
-  })
-  .meta({ id: 'TargetEntry' });
-
 export type KindDescriptor = z.infer<typeof KindDescriptorSchema>;
 export type SourceEntry = z.infer<typeof SourceEntrySchema>;
 export type SourceOrigin = z.infer<typeof SourceOriginSchema>;
-export type TargetEntry = z.infer<typeof TargetEntrySchema>;
-export type TargetVariable = z.infer<typeof TargetVariableSchema>;
 export type TierDescriptor = z.infer<typeof TierDescriptorSchema>;
-export type TokenMapping = z.infer<typeof TokenMappingSchema>;
-export type TokenMappingEntry = z.infer<typeof TokenMappingEntrySchema>;
